feat(medal): add setting and logic to show/hide medal if your pb is better
fix(pb): remove pb from api call, instead keep the one requested before

this is because of unplanned logic in the surround api. this always create a row for the account even if they never played the map, making the returned row "fake" for a specific score if there's already a record at that score.