fix(custom position): fix custom positions resetting when you have none saved
fixes #37
When you decided to have no custom positions, reloading the plugin would cause it to put back the default custom position, since it detected an empty string.
This is fixed by saving "empty" when the user decided to have no custom position, then just load nothing when we detect that string when loading the settings.